% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD_sumstats.R
\name{CFD_sumstats}
\alias{CFD_sumstats}
\title{CFD_sumstats}
\usage{
CFD_sumstats(project)
}
\arguments{
\item{project}{Name of project to include}
}
\description{
Calculates summary statistics from CFD continuous data templates, and creates AWQMS import files. Running function
will bring up a file selctor. Summary stats file will be saved at the same location.
}
