% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD_sumstats 2024BLM.R
\name{CFD_sumstats_2024_BLM}
\alias{CFD_sumstats_2024_BLM}
\title{CFD_sumstats_2024_BLM}
\usage{
CFD_sumstats_2024_BLM(project, path = NULL, is_salem = FALSE)
}
\arguments{
\item{project}{Name of project for AWQMS}

\item{path}{filepath for file to process, if type == "path"}

\item{type}{How to select files. "file" lets you specify path in function arguments, 'file_select' brings up a chooser window, "directory" will loop through all files in a directory.}
}
\description{
Calculates summary statistics from BLM excel exports in the 2024 CFD, and creates AWQMS import files.
Summary stats file will be saved at the same location ans input file.
}
