% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copbes_awqms.R
\name{copbes_AWQMS}
\alias{copbes_AWQMS}
\title{Retrieve surface water monitoring results from The City of Portland Bureau of Environmental Services and creates and
saves AWQMS import files}
\usage{
copbes_AWQMS(data_inventory, project)
}
\arguments{
\item{project}{project name for AWQMS}

\item{station}{Required string vector of station location IDs to be fetched.}

\item{startdate}{Required string setting the start date of the data being fetched. Format 'yyyy-mm-dd'.}

\item{enddate}{Required string setting the end date of the data being fetched. Format 'yyyy-mm-dd'.}

\item{char}{Required string vector identifying the characteristics to be fetched. Options include:
\itemize{
\item 'Dissolved oxygen.Field Visits' - Dissolved oxygen concentration grab measurements (mg/l)
\item 'Dissolved oxygen.Primary' - Continuous dissolved oxygen concentration (mg/l)
\item 'Dissolved oxygen saturation.Field Visits' (\%)
\item 'Dissolved oxygen saturation.Primary' (\%)
\item 'pH.Field Visits' - pH grab measurements (pH)
\item 'pH.Primary' - Continuous pH (pH)
\item 'Specific conductance.Field Visits' - Specific conductance grab measurements (uS/cm)
\item 'Specific conductance.Primary' - Continuous Specific conductance (uS/cm)
\item 'Temperature.Field Visits' - Water temperature grab measurements (deg C)
\item 'Temperature.Primary' - Continuous Water temperature (deg C)
\item 'Temperature.7DADM' - 7-Day mean daily maximum water temperature (deg C)
}}

\item{save_location}{location to save files to. Should end with/}
}
\value{
nothing
}
\description{
Data is retrieved from City of Portland Bureau of Environmental Services database.
https://aquarius.portlandoregon.gov/ using the copbes_data function
}
