% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_check_AWQMS.R
\name{dup_check_AWQMS}
\alias{dup_check_AWQMS}
\title{dup_check_AWQMS}
\usage{
dup_check_AWQMS(
  df,
  mloc_col = "SiteID",
  date_col = "ActStartDate",
  time_col = "ActStartTime",
  char_col = "CharID",
  stat_basis_col = "StatisticalBasis",
  depth_col = "SmplDepth",
  method_col = "Method",
  result_col = "Result"
)
}
\arguments{
\item{df}{Datafram of AWQMS import data to be checked against VW_AWQMS_Results in AWQMS}

\item{mloc_col}{Character of column name that holds MLocID information}

\item{date_col}{Character of column name that holds SampleStartDate information}

\item{time_col}{Character of column name that holds SampleStartTime information}

\item{char_col}{Character of column name that holds Char_Name information}

\item{stat_basis_col}{Character of column name that holds Statistical_Base information}

\item{depth_col}{Character of column name that holds act_depth_height information}

\item{method_col}{Character of column name that holds Method_Code information}

\item{result_col}{Character of column name that holds SampleStartDate information}
}
\description{
This function checks AWQMS for suspected duplicates when generating AWQMS import files. Function will query data
from AWQMS, join to provided datatable, and create a list of 3 objects: non_duplicates, suspected_dups,
suspected_updates. Suspected updates are matches in AWQMS, but with different results, suggesting there has been an
update somewhere (original dataset, data processing, etc..)
}
