% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/National_Estuary_sum_stat.R
\name{NERRS_sum_stats}
\alias{NERRS_sum_stats}
\title{Formats files downloaded from NERRS at http://cdmo.baruch.sc.edu/get/landing.cfm into AWQMS import files}
\usage{
NERRS_sum_stats(path)
}
\description{
Download csv files and  save to a folder. The function will open a directory selector. Point the function to the
directory where the csv files are saved. AWQMS import files will be saved in the same location
}
