% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_air_tmax.R
\name{noaa_air_tmax}
\alias{noaa_air_tmax}
\title{Retrieve historic climate data from NOAA}
\usage{
noaa_air_tmax(station, startdate, enddate, data_type = "TMAX")
}
\arguments{
\item{station}{noaa air temperature station}

\item{startdate}{start date}

\item{enddate}{end date}

\item{data_type}{Data to return. Default is TMAX}
}
\description{
Data is retrieved NOAA's NCEI Data service.
https://www.ncei.noaa.gov/support/access-data-service-api-user-documentation
}
