% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWRD_2_AWQMS.R
\name{owrd_AWQMS}
\alias{owrd_AWQMS}
\title{owrd_data_2_AWQMS
format data from OWRD into an AWQMS format}
\usage{
owrd_AWQMS(df, project = "")
}
\arguments{
\item{df}{dataframe to format into AWQMS}

\item{project}{project to put data into}
}
\value{
data frame of data formatted
}
\description{
This function takes data exported from OWRD using the owrd_data() function and formats for AWQMS. Only high quality,
published temperature data should be passed to this function as it doesn't do any QC checks, nor is it currently
setup for DO or other parameters.
}
