% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owrd_data.R
\name{owrd_data}
\alias{owrd_data}
\title{Retrieve surface water monitoring results from Oregon Water Resources Department}
\usage{
owrd_data(station, startdate, enddate, char)
}
\arguments{
\item{station}{Required string vector of stations to be fetched.}

\item{startdate}{Required string setting the start date of the data being fetched. Format 'yyyy-mm-dd'.}

\item{enddate}{Required string setting the end date of the data being fetched. Format 'yyyy-mm-dd'.}

\item{char}{Required string vector identifying the characteristics to be fetched. Options include:
\itemize{
\item 'MDF' - Mean Daily Flow
\item 'Instantaneous_Stage' - Instantaneous Stage
\item 'Instantaneous_Flow' - Instantaneous Flow
\item 'Measurements' - Discharge grab measurements
\item 'WTEMP_MEASURE' - Water temperature grab measurements
\item 'WTEMP15' - Instantaneous Water temperature
\item 'WTEMP_MEAN' - Daily mean Water temperature
\item 'WTEMP_MAX' - Daily maximum Water temperature
\item 'WTEMP_MIN' - Daily minimum Water temperature
}}
}
\value{
data frame of data
}
\description{
Data is retrieved from Oregon Water Resources Department's near real time monitoring results.
https://apps.wrd.state.or.us/apps/sw/hydro_near_real_time/
}
