% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDX_BES_datapull.R
\name{PDX_BES_data}
\alias{PDX_BES_data}
\title{PDX_BES_cont_data_pull}
\usage{
PDX_BES_data(startdate, enddate, userID, password, save_location)
}
\arguments{
\item{userID}{UserId for Aquarius API. Contact BES}

\item{password}{password for Aquarius API. Contact BES}

\item{save_location}{filepath to save AWQMS import files}
}
\description{
This function queries continuous data from BES's Aquarius database and produces AWQMS import files. Temperature, pH, and DO
are queried and IR summary statistics are calculated. THe function outputs three csv files: Summary Stats,
continuous pH, and monitoring location information.
}
