% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copbes_awqms.R
\name{copbes_AWQMS}
\alias{copbes_AWQMS}
\title{Retrieve surface water monitoring results from The City of Portland Bureau of Environmental Services and creates and
saves AWQMS import files}
\usage{
copbes_AWQMS(data_inventory, project)
}
\arguments{
\item{data_inventory}{Required datafarme of available data. Columns should match augments of the copbes_data function:
\itemize{
\item 'station'
\item 'startdate'
\item 'enddate'
\item 'char'
}}

\item{project}{project name for AWQMS}
}
\value{
list of deployments, sumstats, and raw pH continuous data
}
\description{
Data is retrieved from City of Portland Bureau of Environmental Services database.
https://aquarius.portlandoregon.gov/ using the copbes_data function
}
